pyproject.toml
[project]
name = "bocek"
version = "0.1.0"
description = "Discord bot to keep you company."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "discord-py[voice]>=2.4.0",
    "google-api-core==2.9.0",
    "google-api-python-client>=2.154.0",
    "google-cloud-texttospeech>=2.12.3",
    "google-cloud>=0.34.0",
    "gtts>=2.5.4",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "unidecode>=1.3.8",
    "deepdiff>=8.0.1",
    "pynacl>=1.5.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.62.3",
    "aiohttp>=3.11.7",
    "aiohttp-retry>=2.9.1",
    "async-property>=0.2.2",
    "mutagen>=1.47.0",
    "niquests>=3.12.1",
]

[dependency-groups]
dev = ["ruff>=0.8.0"]

[tool.ruff]
extend-select = ["I"]
select = ["E4", "E7", "E9", "F", "B"]

[tool.bocek.rito]
lol-game-ip = "192.168.0.31"
lol-game-port = 29999
offline-wait = 30
online-wait = 2
players = [
    "Ciszkoo",
    "LikeBanana",
    "Chonkey",
    "SwagettiYoloneze",
    "Sabijak",
    "Xubeks",
    "Nowik6300",
    "GodRevi",
    "Mαster Vi",
]

[tool.bocek.rito.event-possibility]
PentaKill = 1
QuadraKill = 0.8
TripleKill = 0.5
BaronSteal = 1
DragonSteal = 0.8
DoubleKill = 0.4
BaronKill = 0.5
Ace = 0.5
DragonKill = 0.3
HeraldKill = 0.3
AtakhanKill = 0.9
FirstBlood = 0.5
ChampionKill = 0.3
ChampionDeath = 0.3
InhibKilled = 0.3
TurretKilled = 0.3
